fix: add pin icons and category icon and restructure in folder components
